UserStatus component
