+++ web/src/components/FaceImages.css
@@ -31,6 +31,7 @@
   font-size: 0.5rem;
   line-height: 0.5rem;
   font-family: sans-serif;
+  cursor: help;
 }
 
 .FaceImages-live-indicator {
