+++ web/src/components/FaceImages.tsx
@@ -39,7 +39,12 @@ const FaceImage = React.memo<{
         />
       )}
       <div className="FaceImages-name">{nickname}</div>
-      <div className="FaceImages-mesg">{statusMesg}</div>
+      <div
+        className="FaceImages-mesg"
+        title={[...statusMesg][1] ? statusMesg : "Enter status message"}
+      >
+        {[...statusMesg][0]}
+      </div>
       {liveMode && !obsoleted && stream && (
         <div className="FaceImages-live-indicator" title="Live Mode On">
           &#9673;
